gmail-agent-project/requirements.txt
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.35.1
google-cloud-aiplatform
google-cloud-functions